YUK-10542, CMEP file format support  - Changed to use PersistedSystemValue table that already exists to keep track of the last changeId possibly used to generate the CMEP file format.  - Added new PersistedSystemValue enum for CMEP format.  - Deleted RPHServiceTag dao and pojo.  - Added token arg to BillingFile.main for use from MACS batch file.  - Changed RPHDao uses of changeId to be long instead of int.  - Added support for token being passed through macs (init.tcl support).

+++ yukon-client/common/src/com/cannontech/core/dao/RawPointHistoryDao.java
@@ -140,7 +140,7 @@ public interface RawPointHistoryDao {
      * @param order - controls ordering by timestamp (only affects the iteration order of the values)
      * @return
      */
-    public ListMultimap<PaoIdentifier, PointValueQualityHolder> getAttributeData(Iterable <? extends YukonPao> paos, Attribute attribute, final int startChangeId, final int stopChangeId, final boolean excludeDisabledPaos, final Clusivity clusivity, final Order order);
+    public ListMultimap<PaoIdentifier, PointValueQualityHolder> getAttributeData(Iterable <? extends YukonPao> paos, Attribute attribute, final long startChangeId, final long stopChangeId, final boolean excludeDisabledPaos, final Clusivity clusivity, final Order order);
     
     /**
      * This method returns RawPointHistory data for a list of PAOs and a given Attribute. This data will be returned as a ListMultimap
@@ -284,7 +284,7 @@ public interface RawPointHistoryDao {
    /**
     * Get the largest changeId currently in RawPointHistory
     */
-   public int getMaxChangeId();
+   public long getMaxChangeId();
    
    public static class AdjacentPointValues {
        private PointValueHolder preceding;
